Fixed topology optimization

+++ ALSolver/Solver.cpp
@@ -2,19 +2,19 @@
 
 #include <unsupported/Eigen/SparseExtra>
 
-int SolveFE(int* rows_offset, int* cols, float* vals, float* F, int dim, int dof, int nnz, float* X)
+int SolveFE(int* rows_offset, int* cols, double* vals, double* F, int dim, int dof, int nnz, double* X)
 {
-    Eigen::Map<Eigen::SparseMatrix<float, Eigen::StorageOptions::RowMajor>> A(dim, dim, nnz, rows_offset, cols, vals);
+    Eigen::Map<Eigen::SparseMatrix<double, Eigen::StorageOptions::RowMajor>> A(dim, dim, nnz, rows_offset, cols, vals);
     auto B = SetVector(F, dim);
 
     //Eigen::saveMarket(A, "C:/Users/alber/Desktop/mat.mtx", Eigen::UpLoType::Symmetric);
     //Eigen::saveMarket(B, "C:/Users/alber/Desktop/B.mtx");
 
-    Eigen::VectorXf result;
+    Eigen::VectorXd result;
 
     if (dim < 20000)
     {
-        Eigen::SimplicialLLT<Eigen::SparseMatrix<float>> llt(A);
+        Eigen::SimplicialLLT<Eigen::SparseMatrix<double>> llt(A);
 
         llt.analyzePattern(A);
         llt.factorize(A);
@@ -23,7 +23,7 @@ int SolveFE(int* rows_offset, int* cols, float* vals, float* F, int dim, int dof
     }
     else
     {
-        Eigen::PardisoLLT<Eigen::SparseMatrix<float>, 1> pardiso;
+        Eigen::PardisoLLT<Eigen::SparseMatrix<double>, 1> pardiso;
         //pardiso.pardisoParameterArray()[3] = 32; // PCG method
         pardiso.pardisoParameterArray()[59] = 0;
         pardiso.pardisoParameterArray()[1] =3; // 	The parallel (OpenMP) version of the nested dissection algorithm.
