+++ ALSolver/Solver.h
@@ -7,6 +7,6 @@
 #include <Eigen/Eigen>
 #include <Eigen/PardisoSupport>
 
-Eigen::VectorXf SetVector(float* matrix, int dim);
-Eigen::SparseMatrix<float> SetCOOMatrix(int* row, int* col, float* val, int rows, int cols, int nnz);
-extern "C" __declspec(dllexport) int SolveFE(int* rows_offset, int* cols, float* vals, float* F, int dim, int dof, int nnz, float* X);
+Eigen::VectorXd SetVector(double* matrix, int dim);
+Eigen::SparseMatrix<double> SetCOOMatrix(int* row, int* col, double* val, int rows, int cols, int nnz);
+extern "C" __declspec(dllexport) int SolveFE(int* rows_offset, int* cols, double* vals, double* F, int dim, int dof, int nnz, double* X);
