+++ FEModel/Elements/Pixel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MathNet.Numerics.LinearAlgebra.Single;
+using MathNet.Numerics.LinearAlgebra.Double;
 
 namespace ALFE
 {
@@ -32,12 +32,12 @@ namespace ALFE
         /// </summary>
         public override void ComputeKe()
         {
-            float coeff = Material.E / (1.0f - Material.nu * Material.nu);
+            double coeff = Material.E / (1.0 - Material.nu * Material.nu);
 
-            float[] array = new float[8]
+            double[] array = new double[8]
             {
-                0.5f-Material.nu/6.0f , 0.125f + Material.nu * 0.125f, -0.25f - Material.nu / 12.0f, -0.125f + Material.nu * 0.375f,
-                -0.25f + Material.nu/12f, -0.125f - Material.nu *0.125f, Material.nu / 6.0f, 0.125f- Material.nu*0.375f
+                0.5-Material.nu/6.0 , 0.125f + Material.nu * 0.125f, -0.25f - Material.nu / 12.0, -0.125f + Material.nu * 0.375f,
+                -0.25f + Material.nu/12f, -0.125f - Material.nu *0.125f, Material.nu / 6.0, 0.125f- Material.nu*0.375f
              };
 
             for (int i = 0; i < 8; i++)
@@ -45,7 +45,7 @@ namespace ALFE
                 array[i] *= coeff;
             }
 
-            Ke = DenseMatrix.OfArray(new float[8, 8]
+            Ke = DenseMatrix.OfArray(new double[8, 8]
             {
                 { array[0], array[1], array[2],array[3],array[4],array[5], array[6], array[7]},
                 { array[1], array[0], array[7],array[6],array[5],array[4], array[3], array[2]},
