+++ FEModel/Elements/Triangle.cs
@@ -3,16 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MathNet.Numerics.LinearAlgebra.Single;
+using MathNet.Numerics.LinearAlgebra.Double;
 
 namespace ALFE
 {
     public class Triangle : Element
     {
-        public float Area;
-        public float Thickness = 1.0f;
+        public double Area;
+        public double Thickness = 1.0;
 
-        public Triangle(List<Node> nodes, Material material, float thickness = 1.0f, bool exist = true)
+        public Triangle(List<Node> nodes, Material material, double thickness = 1.0, bool exist = true)
         {
             if (nodes.Count != 3)
                 throw new Exception("The number of nodes must be 3.");
@@ -36,16 +36,16 @@ namespace ALFE
         {
             D = new DenseMatrix(3, 3);
 
-            float coeff1 = Material.E / (1.0f - Material.nu* Material.nu);
+            double coeff1 = Material.E / (1.0 - Material.nu* Material.nu);
 
             D[0, 0] = D[1, 1] = coeff1;
             D[0, 1] = D[1, 0] = Material.nu * coeff1;
-            D[2, 2] = (1.0f - Material.nu) * 0.5f * coeff1;
+            D[2, 2] = (1.0 - Material.nu) * 0.5 * coeff1;
         }
 
         public void ComputeB()
         {
-            var val = 1.0f / (2.0f * Area);
+            var val = 1.0 / (2.0 * Area);
 
             var belta0 = (Nodes[1].Position.Y - Nodes[2].Position.Y) * val;
             var belta1 = (Nodes[2].Position.Y - Nodes[0].Position.Y) * val;
@@ -55,10 +55,10 @@ namespace ALFE
             var gama1 = (Nodes[0].Position.X - Nodes[2].Position.X) * val;
             var gama2 = (Nodes[1].Position.X - Nodes[0].Position.X) * val;
 
-            B = DenseMatrix.OfArray(new float[3,6]
+            B = DenseMatrix.OfArray(new double[3,6]
                {
-                    { belta0, 0.0f, belta1, 0.0f, belta2, 0.0f },
-                    { 0.0f, gama0, 0.0f, gama1, 0.0f, gama2 },
+                    { belta0, 0.0, belta1, 0.0, belta2, 0.0 },
+                    { 0.0, gama0, 0.0, gama1, 0.0, gama2 },
                     { gama0, belta0, gama1, belta1, gama2,belta2 }
                });
         }
@@ -78,7 +78,7 @@ namespace ALFE
         {
             Area = Math.Abs(Nodes[0].Position.X * (Nodes[1].Position.Y - Nodes[2].Position.Y) +
                 Nodes[1].Position.X * (Nodes[2].Position.Y - Nodes[0].Position.Y) + 
-                Nodes[2].Position.X * (Nodes[0].Position.Y - Nodes[1].Position.Y)) * 0.5f;
+                Nodes[2].Position.X * (Nodes[0].Position.Y - Nodes[1].Position.Y)) * 0.5;
         }
     }
 }
