+++ FEModel/Material.cs
@@ -5,18 +5,18 @@
         /// <summary>
         /// // Young's modulus
         /// </summary>
-        public float E { get; set; }
+        public double E { get; set; }
 
         /// <summary>
         /// Poisson ratio
         /// </summary>
-        public float nu { get; set; }
+        public double nu { get; set; }
         public Material() 
         {
-            this.E = 1.0f;
+            this.E = 1.0;
             this.nu = 0.3f;
         }
-        public Material(float E, float nu)
+        public Material(double E, double nu)
         {
             this.E = E;
             this.nu = nu;
